requirements.txt
alembic==1.3.0
astroid==2.3.3
autopep8==1.4.4
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
entrypoints==0.3
firebase-admin==3.1.0
flake8==3.7.9
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.6.0
google-cloud-storage==1.22.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
pg8000==1.13.2
protobuf==3.10.0
psycopg2-binary==2.8.4
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
rsa==4.0
scramp==1.1.0
six==1.13.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
